requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
filelock==3.18.0
fish-audio-sdk==2025.4.2
frozenlist==1.6.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
idna==3.10
jiter==0.9.0
multidict==6.4.3
openai==1.77.0
ormsgpack==1.9.1
propcache==0.3.1
protobuf==5.29.4
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyht==0.1.14
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==13.1
webvtt-py==0.5.1
wsproto==1.2.0
yarl==1.20.0
